배경택:[Add] 환경설정 관련(화면모드, 해상도, 밝기, 조준점, 텍스트, 사운드) 스크립트 추가
