+++ Metalord_btin/MetaLord/Assets/_Test/BKT/Scripts/ReadyMenu/Settings/BrightnessSetting.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6c1b654e407cd49798fd8bf2a43d7efc
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
