+++ Metalord_btin/MetaLord/Assets/_Test/BKT/Scripts/ReadyMenu/Settings/CrossLineSetting.cs
@@ -1,0 +1,18 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class CrossLineSetting : MonoBehaviour
+{
+    // Start is called before the first frame update
+    void Start()
+    {
+        
+    }
+
+    // Update is called once per frame
+    void Update()
+    {
+        
+    }
+}
