+++ Metalord_btin/MetaLord/Assets/_Test/BKT/Scripts/ReadyMenu/Settings/CrossLineSetting.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: fae9b578758bd477894402fd2f4e3e97
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
