+++ Metalord_btin/MetaLord/Assets/_Test/BKT/Scripts/ReadyMenu/Settings/DisplaySetting.cs
@@ -1,0 +1,12 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+/// <summary>
+/// 디스플레이 셋팅(전체화면, 창화면)
+/// 240102 배경택
+/// </summary>
+public class DisplaySetting : MonoBehaviour
+{
+    
+}
