+++ Metalord_btin/MetaLord/Assets/_Test/BKT/Scripts/ReadyMenu/Settings/DisplaySetting.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1a46dcb85d00c432486fbc50be227ccd
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
