+++ Metalord_btin/MetaLord/Assets/_Test/BKT/Scripts/ReadyMenu/Settings/KeyRebind.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 15524f5390bcf404fbfd28b52e604d58
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
