+++ Metalord_btin/MetaLord/Assets/_Test/BKT/Scripts/ReadyMenu/Settings/KeyRebind/RebindKeyUI.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1184001f4c3fb4455ae40e9e662dc4ec
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
