+++ Metalord_btin/MetaLord/Assets/_Test/BKT/Scripts/ReadyMenu/Settings/ResolutionSetting.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e63ae485544d643b7bec36249b62b3dd
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
