+++ Metalord_btin/MetaLord/Assets/_Test/BKT/Scripts/ReadyMenu/Settings/SettingsCanvas.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 975df7fd96f4e442eb401cc23f2c3659
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
