+++ Metalord_btin/MetaLord/Assets/_Test/BKT/Scripts/ReadyMenu/Settings/TextSizeSetting.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ffbade7c1ead14f81ae1d14f193edc64
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
